fix: prevent surrogate pairs from being split by the selected range
This is already handled in functions like `expandSelectionInDirection`
which won't select a half of emoji etc. but always the whole surrogate
pair. The solution here is the same, if the cursor is placed in the
middle of a surrogate pair, it will be moved to the end of the pair.
Similarly a selection will either select both or neither.